scraper/requirements.txt
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
filelock==3.12.4
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.8.1
Protego==0.3.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.10.1
service-identity==23.1.0
soupsieve==2.5
tldextract==3.5.0
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==2.0.4
w3lib==2.1.2
zope.interface==6.0
